pyproject.toml
[tool.poetry]
name = "dan-toropov-cw-7"
version = "0.1.0"
description = ""
authors = ["dan.toropov <dan.toropov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
pillow = "^11.0.0"
psycopg-binary = "^3.2.3"
dotenv-python = "^0.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
black = "^24.10.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
django-apscheduler = "^0.7.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
